Remove reallyBuildDroid and refactor uint, UDWORD etc. to int reallyBuildDroid was removed because it was superfluous. In the background the various datatypes such as UDWORD, UBYTE etc. are refactored to int This is because the value-range of int is for all applications enough except in very rare cases. Additionally ints allow easier "hacks" with uninitialized data and various extra states. E.g. formerly constructs such as: group = UDWORD_MAX; existed, they will be replaced by group = -1;

+++ src/structure.cpp
@@ -2341,7 +2341,8 @@ static bool structPlaceDroid(STRUCTURE *psStructure, DROID_TEMPLATE *psTempl, DR
 
 	if (placed)
 	{
-		INITIAL_DROID_ORDERS initialOrders = {psStructure->pFunctionality->factory.secondaryOrder, psStructure->pFunctionality->factory.psAssemblyPoint->coords.x, psStructure->pFunctionality->factory.psAssemblyPoint->coords.y, psStructure->id};
+		//FIXME where is FUNCTIONALITY defined? Can't seem to find it... Either way, this cast seems dangerous
+		INITIAL_DROID_ORDERS initialOrders = {(int) psStructure->pFunctionality->factory.secondaryOrder, psStructure->pFunctionality->factory.psAssemblyPoint->coords.x, psStructure->pFunctionality->factory.psAssemblyPoint->coords.y, psStructure->id};
 		//create a droid near to the structure
 		syncDebug("Placing new droid at (%d,%d)", x, y);
 		turnOffMultiMsg(true);
@@ -3948,7 +3949,7 @@ void structureUpdate(STRUCTURE *psBuilding, bool mission)
 	CHECK_STRUCTURE(psBuilding);
 }
 
-STRUCTURE::STRUCTURE(uint32_t id, unsigned player)
+STRUCTURE::STRUCTURE(int id, int player)
 	: BASE_OBJECT(OBJ_STRUCTURE, id, player)
 	, pFunctionality(nullptr)
 	, buildRate(1)  // Initialise to 1 instead of 0, to make sure we don't get destroyed first tick due to inactivity.
